backend/fastapi_service/pyproject.toml
[project]
name = "ansflow-fastapi-service"
version = "1.0.0"
description = "AnsFlow CI/CD Platform - FastAPI High-Performance Service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # FastAPI 核心框架和服务器
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    # 数据库和ORM
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0", # PostgreSQL异步驱动
    "aiomysql>=0.2.0", # MySQL异步驱动
    "alembic>=1.12.0", # 数据库迁移
    # 数据验证和序列化
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    # 异步HTTP客户端
    "httpx>=0.25.0",
    "aiofiles>=23.2.1",
    # WebSocket支持
    "websockets>=12.0",
    # 消息队列客户端
    "aio-pika>=9.3.0", # RabbitMQ异步客户端
    "redis>=5.0.0", # Redis客户端
    # 认证和安全
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    # 日志和监控
    "structlog>=23.2.0",
    "prometheus-client>=0.19.0",
    # 配置管理
    "python-dotenv>=1.0.0",
    # 工具和实用库
    "python-json-logger>=2.0.7",
    "typer>=0.9.0", # CLI工具
    # 开发和测试
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0", # 测试HTTP客户端
    "email-validator>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ansflow_api"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["ansflow_api"]
